requirements.txt
torch>=1.3.0
torchvision>=0.4.1
tensorboard>=2.3.0
Pillow>=6.2.1
